+++ zcash_client_sqlite/CHANGELOG.md
@@ -10,6 +10,12 @@ and this library adheres to Rust's notion of
 ### Added
 - A new `orchard` feature flag has been added to make it possible to
   build client code without `orchard` dependendencies.
+- `impl From<zcash_keys::keys::AddressGenerationError> for SqliteClientError`
+
+### Changed
+- `zcash_client_sqlite::error::SqliteClientError` has changed variants:
+  - Added `AddressGeneration`
+  - Removed `DiversifierIndexOutOfRange`
 
 ## [0.9.0] - 2024-03-01